+++ src/main/java/net/kline72/ksnrpgmod/util/MobLevelUtil.java
@@ -22,20 +22,20 @@ public class MobLevelUtil {
 
         LevelAccessor world = event.getLevel();
         double distance = 0;
-        double mLevel;
+        int mLevel;
 
         if (!entity.getPersistentData().getBoolean("scaled")) {
             distance = Math.sqrt(Math.pow(entity.getX() - world.getLevelData().getXSpawn(), 2) + Math.pow(entity.getZ() - world.getLevelData().getZSpawn(), 2));
             mLevel = (int) Math.round(distance / 100);
-            entity.getPersistentData().putDouble("mobLevel", mLevel);
+            entity.getPersistentData().putInt("mobLevel", mLevel);
             if (event.getEntity() instanceof Monster) {
-                entity.getAttribute(Attributes.MAX_HEALTH).setBaseValue(entity.getAttributeBaseValue(Attributes.MAX_HEALTH) + entity.getPersistentData().getDouble("mobLevel"));
-                entity.getAttribute(Attributes.ATTACK_DAMAGE).setBaseValue(entity.getAttributeBaseValue(Attributes.ATTACK_DAMAGE) + (entity.getPersistentData().getDouble("mobLevel") * 0.2));
+                entity.getAttribute(Attributes.MAX_HEALTH).setBaseValue(entity.getAttributeBaseValue(Attributes.MAX_HEALTH) + entity.getPersistentData().getInt("mobLevel"));
+                entity.getAttribute(Attributes.ATTACK_DAMAGE).setBaseValue(entity.getAttributeBaseValue(Attributes.ATTACK_DAMAGE) + (entity.getPersistentData().getInt("mobLevel") * 0.2));
             } else if (event.getEntity() instanceof Animal || event.getEntity() instanceof Villager) {
-                entity.getAttribute(Attributes.MAX_HEALTH).setBaseValue(entity.getAttributeBaseValue(Attributes.MAX_HEALTH) + (entity.getPersistentData().getDouble("mobLevel") * 2));
+                entity.getAttribute(Attributes.MAX_HEALTH).setBaseValue(entity.getAttributeBaseValue(Attributes.MAX_HEALTH) + (entity.getPersistentData().getInt("mobLevel") * 2));
             } else if (event.getEntity() instanceof IronGolem) {
-                entity.getAttribute(Attributes.MAX_HEALTH).setBaseValue(entity.getAttributeBaseValue(Attributes.MAX_HEALTH) + (entity.getPersistentData().getDouble("mobLevel") * 2));
-                entity.getAttribute(Attributes.ATTACK_DAMAGE).setBaseValue(entity.getAttributeBaseValue(Attributes.ATTACK_DAMAGE) + (entity.getPersistentData().getDouble("mobLevel") * 0.4));
+                entity.getAttribute(Attributes.MAX_HEALTH).setBaseValue(entity.getAttributeBaseValue(Attributes.MAX_HEALTH) + (entity.getPersistentData().getInt("mobLevel") * 2));
+                entity.getAttribute(Attributes.ATTACK_DAMAGE).setBaseValue(entity.getAttributeBaseValue(Attributes.ATTACK_DAMAGE) + (entity.getPersistentData().getInt("mobLevel") * 0.4));
             }
             entity.getPersistentData().putBoolean("scaled", true);
             entity.setHealth(entity.getMaxHealth());
